DOMA-3099 extracted tasks tracking components into 'tasks' folder
Rebase after @sitozzz has copypasted all created files in this commit.
